Move sync / rollback method to the separate file
